Add TODO comments
Adding my thoughts on the algorithm for checking if file already exists at destination.

This can absolutely happen if someone were to write files to the slower tiers instead of the top level tier, because the top level tier doesnt have enough space to write to. I have already run into this issue, and I get around it by having a samba share pointing to the larger, slower tier. This can be helpful to avoid writing tons of data to SSDs that you know will be cold data for most of its life, and the spinning HDDs should be able to handle large sequential file transfers just fine on its own.